Script to create or update index.md files (#32416)
